ボタン連打時のエラー修正
