+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { Header } from "./header"
 export { Footer } from "./footer"
+export { PrefPopulation } from "./prefPopulation"
